ai_backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
cd